8316000: File.setExecutable silently fails if file does not exist

+++ test/jdk/java/io/File/SetAccess.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2005, 2012, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2005, 2023, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -22,15 +22,21 @@
  */
 
 /* @test
-   @bug 4167472 5097703 6216563 6284003 6728842 6464744
-   @summary Basic test for setWritable/Readable/Executable methods
-   @build SetAccess Util
-   @run main SetAccess
+ * @bug 4167472 5097703 6216563 6284003 6728842 6464744 8316000
+ * @summary Basic test for setWritable/Readable/Executable methods
+ * @library .. /test/lib
+ * @build jdk.test.lib.Platform
+ * @build SetAccess Util
+ * @run main SetAccess
  */
 
-import java.io.*;
-import java.nio.file.*;
-import java.nio.file.attribute.*;
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.OutputStream;
+import java.nio.file.Files;
+import java.nio.file.attribute.PosixFileAttributes;
+import java.nio.file.attribute.PosixFilePermissions;
+import jdk.test.lib.Platform;
 
 public class SetAccess {
     public static void main(String[] args) throws Exception {
@@ -49,18 +55,36 @@ public class SetAccess {
         o.close();
         doTest(f);
 
+        if (f.exists() && !f.delete())
+            throw new Exception("Can't delete test file: " + f);
+
         f = new File(d, "x.SetAccessPermission.dir");
         if (f.exists() && !f.delete())
             throw new Exception("Can't delete test dir: " + f);
         if (!f.mkdir())
             throw new Exception(f + ": Cannot create directory");
         doTest(f);
+
+        if (f.exists() && !f.delete())
+            throw new Exception("Can't delete test dir: " + f);
     }
 
     public static void doTest(File f) throws Exception {
-        if (!System.getProperty("os.name").startsWith("Windows")) {
+        // Generic
+        if (f.isFile()) {
+            File no = new File("nonexistent");
+            assert !no.exists();
+            if (no.setReadable(true, false))
+                throw new Exception(no + ": setReadable(true, false) Succeeded");
+            if (no.setWritable(true, false))
+                throw new Exception(no + ": setWritable(true, false) Succeeded");
+            if (no.setExecutable(true, false))
+                throw new Exception(no + ": setExecutable(true, false) Succeeded");
+        }
+
+        if (!Platform.isWindows()) {
             if (!f.setReadOnly())
-                 throw new Exception(f + ": setReadOnly Failed");
+                throw new Exception(f + ": setReadOnly Failed");
             if (!f.setWritable(true, true) ||
                 !f.canWrite() ||
                 permission(f).charAt(2) != 'w')
@@ -127,8 +151,7 @@ public class SetAccess {
             if (!f.setReadable(false) || f.canRead() ||
                 permission(f).charAt(1) != '-')
                 throw new Exception(f + ": setReadable(false, true) Failed");
-        } else {
-            //Windows platform
+        } else { // Windows platform
             if (f.isFile()) {
                 if (!f.setReadOnly())
                     throw new Exception(f + ": setReadOnly Failed");
@@ -150,39 +173,37 @@ public class SetAccess {
                     throw new Exception(f + ": setReadable(true, false) Failed");
                 if (!f.setReadable(true) || !f.canRead())
                     throw new Exception(f + ": setReadable(true, true) Failed");
+                if (f.isDirectory()) {
+                    // setWritable should fail on directories because the DOS readonly
+                    // attribute prevents a directory from being deleted.
+                    if (f.setWritable(false, true))
+                        throw new Exception(f + ": setWritable(false, true) Succeeded");
+                    if (f.setWritable(false, false))
+                        throw new Exception(f + ": setWritable(false, false) Succeeded");
+                    if (f.setWritable(false))
+                        throw new Exception(f + ": setWritable(false) Succeeded");
+                } else {
+                    if (!f.setWritable(false, true) || f.canWrite())
+                        throw new Exception(f + ": setWritable(false, true) Failed");
+                    if (!f.setWritable(false, false) || f.canWrite())
+                        throw new Exception(f + ": setWritable(false, false) Failed");
+                    if (!f.setWritable(false) || f.canWrite())
+                        throw new Exception(f + ": setWritable(false) Failed");
+                }
+                if (f.setExecutable(false, true))
+                    throw new Exception(f + ": setExecutable(false, true) Failed");
+                if (f.setExecutable(false, false))
+                    throw new Exception(f + ": setExecutable(false, false) Failed");
+                if (f.setExecutable(false))
+                    throw new Exception(f + ": setExecutable(false, true) Failed");
+                if (f.setReadable(false, true))
+                    throw new Exception(f + ": setReadable(false, true) Failed");
+                if (f.setReadable(false, false))
+                    throw new Exception(f + ": setReadable(false, false) Failed");
+                if (f.setReadable(false))
+                    throw new Exception(f + ": setReadable(false, true) Failed");
             }
-            if (f.isDirectory()) {
-                // setWritable should fail on directories because the DOS readonly
-                // attribute prevents a directory from being deleted.
-                if (f.setWritable(false, true))
-                    throw new Exception(f + ": setWritable(false, true) Succeeded");
-                if (f.setWritable(false, false))
-                    throw new Exception(f + ": setWritable(false, false) Succeeded");
-                if (f.setWritable(false))
-                    throw new Exception(f + ": setWritable(false) Succeeded");
-            } else {
-                if (!f.setWritable(false, true) || f.canWrite())
-                    throw new Exception(f + ": setWritable(false, true) Failed");
-                if (!f.setWritable(false, false) || f.canWrite())
-                    throw new Exception(f + ": setWritable(false, false) Failed");
-                if (!f.setWritable(false) || f.canWrite())
-                    throw new Exception(f + ": setWritable(false) Failed");
-            }
-            if (f.setExecutable(false, true))
-                throw new Exception(f + ": setExecutable(false, true) Failed");
-            if (f.setExecutable(false, false))
-                throw new Exception(f + ": setExecutable(false, false) Failed");
-            if (f.setExecutable(false))
-                throw new Exception(f + ": setExecutable(false, true) Failed");
-            if (f.setReadable(false, true))
-                throw new Exception(f + ": setReadable(false, true) Failed");
-            if (f.setReadable(false, false))
-                throw new Exception(f + ": setReadable(false, false) Failed");
-            if (f.setReadable(false))
-                throw new Exception(f + ": setReadable(false, true) Failed");
         }
-        if (f.exists() && !f.delete())
-            throw new Exception("Can't delete test dir: " + f);
     }
 
     private static String permission(File f) throws Exception {
